+++ speech_to_text_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.4.2
+- bug fix for null options
+
+## 1.4.1
+- uses platform options in method channel call
+
+## 1.4.0
+- platform specific configuration properties added to initialize
+
 ## 1.3.0
 - listen changed to bool